chore(components): unify default styles types and values, unify Loader component
