--- packages/ui-kit/src/components/Counter/defaults.ts
@@ -1,12 +1,0 @@
-export const defaultStyles = {
-  font: {
-    color: '#000',
-    size: 'inherit',
-    weight: 'inherit',
-    style: 'inherit',
-    family: 'inherit',
-    stretch: 'inherit',
-    variant: 'inherit',
-    lineHeight: 'inherit'
-  }
-}
