+++ packages/ui-kit/src/components/Counter/index.ts
@@ -1,2 +1,1 @@
-import { Container } from './Container'
-export const Counter = Container
+export { Container as Counter } from './Container'
