--- packages/ui-kit/src/components/Counter/types.ts
@@ -1,12 +1,0 @@
-export type Styles = {
-  font?: {
-    color?: string
-    size?: string
-    weight?: string
-    style?: string
-    family?: string
-    stretch?: string
-    variant?: string
-    lineHeight?: string
-  }
-}
