+++ packages/ui-kit/src/components/Leaderboard/Leaderboard.types.ts
@@ -1,0 +1,6 @@
+export type LeaderboardChartVariant = 'bar' | 'table'
+
+export type LeaderboardData = {
+  headers?: string[]
+  rows?: Array<Array<string | null>>
+}
