+++ packages/ui-kit/src/components/Leaderboard/ValueBar/index.ts
@@ -1,0 +1,1 @@
+export * from './ValueBar'
