+++ packages/ui-kit/src/components/TimeSeries/TimeSeries.types.ts
@@ -1,0 +1,11 @@
+import type { ScaleOptionsByType } from 'chart.js'
+import { DeepPartial } from 'chart.js/dist/types/utils'
+
+export type ChartScales = DeepPartial<{ [key: string]: ScaleOptionsByType<'linear' | 'logarithmic'> }>
+
+export type TimeSeriesChartVariant = 'bar' | 'line'
+
+export type TimeSeriesData = {
+  values?: Array<number | null>
+  labels?: string[]
+}
