+++ packages/ui-kit/src/components/index.ts
@@ -1,5 +1,6 @@
 export * from './Counter'
 export * from './ErrorBoundary'
 export * from './Leaderboard'
+export * from './Loader'
 export * from './TimeSeries'
 export * from './withErrorBoundary'
