+++ packages/ui-kit/src/themes/index.ts
@@ -1,0 +1,2 @@
+export * from './default.types'
+export * from './defaultStyles'
